Move FacebookBot into a separete file
